+++ src/common/schema/constants.ts
@@ -1,0 +1,2 @@
+export const NO_USER_ID_ERROR =
+  "Error: Specify the userId or set the DESTINATION_USER_ID in the environment variables of this MCP Server.";
