+++ src/common/schema/textMessage.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+export const textMessageSchema = z.object({
+  type: z.literal("text").default("text"),
+  text: z
+    .string()
+    .max(5000)
+    .describe("The plain text content to send to the user."),
+});
